chore(test): remove sugar.rs
