+++ test/src/specs/sync/invalid_locator_size.rs
@@ -1,4 +1,4 @@
-use crate::util::sugar::out_ibd_mode;
+use crate::util::mining::out_ibd_mode;
 use crate::utils::wait_until;
 use crate::{Net, Node, Spec};
 use ckb_network::SupportProtocols;
