+++ test/src/util/mod.rs
@@ -2,4 +2,3 @@ pub mod chain;
 pub mod check;
 pub mod log_monitor;
 pub mod mining;
-pub mod sugar;
